Update patch set 1
Patch Set 1:

I hope I'm typing this in the right place.

The function declaration says 'key', the prototype says 'value' and that is an accident. I tried to mimic the X25519_public_from_private() which is whether the word value came from. 

I'll try pushing a patch with the correction (and the blank line removed) when I finish typing this. 

As for my use case... I only store the 32 byte private key in my data format... so I needed an easy way to go from private to public.

The 64 byte "private key" is byproduct of the reference implementation returning "private keys" that way. If you look at the code for ED25519_keypair(), it was already using memmove to account for consumers who passed in a single 64 byte buffer, where pub simply pointed to second half of private.

Patch-set: 1