Update patch set 1
Patch Set 1:

> > The 64 byte "private key" is byproduct of the reference
 > > implementation returning "private keys" that way. If you look at
 > > the code for ED25519_keypair(), it was already using memmove to
 > > account for consumers who passed in a single 64 byte buffer,
 > where
 > > pub simply pointed to second half of private.
 > 
 > The private key contains a copy of the public key because the
 > signing operation needs to know the public key and it would be too
 > slow to compute it.
 > 
 > what about:
 > 
 > ED25519_keypair_from_raw_private_key(uint8_t out_public_key[32],
 > uint8_t out_private_key[64], const uint8_t raw_private_key[32])?

I calculate the public key and pass them both in for signing, so I don't have a problem with your suggestion. My only question is whether the X25519_* functions match? They pass around 32 byte private keys, and the public_from_private prototype I created was a direct copy of X25519_public_from_private(). People might not realize the distinction between what's needed for DH and a signature.

The other possible "gotcha" is that inevitably someone will end up writing code which leaks the private portion, because they don't understand the distinction/basis for a keypair, versus a key. But that's a minor issue. 

If you want to go down the keypair road, might I suggest always referring to the 64 byte private key as a "keypair" and the 32 byte variant as a "seed" value instead? Might make it slight less confusing.

Patch-set: 1