Update patch set 1
Patch Set 1:

For me, the important thing is being able to derive the public portion based on the private portion. I like the properties associated with a system that handles the two pieces separately. 

I just think we need to be careful with the wording, since there isn't an established vocabulary for ed25519 keys. 

Personally I think of the 32 byte priv key value as being akin to a compressed public key point on a SEC 1 curve, that has been  serialized according X9.62. Those keys need to be "unpacked" before they can be used, just like a 32 byte ed25519 priv key needs to be unpacked before it can be used for signing.

That said, we've beaten this horse well past its death, and since I prefer to avoid animal cruelty, I shall defer to the judgment of the maintainer as to the API, and adapt accordingly.

Patch-set: 1