Update patch set 3
Patch Set 3:

As far as use cases go: I have defined this function in *ring* I use it to check consistency between the public and private parts of a stored key, by doing the derivation and then verifying that the the derived public value matches the stored public value. IIRC, this is consistent with similar checks done in BoringSSL's RSA and ECC private key deserialization code.

Patch-set: 3