Update patch set 1
Patch Set 1:

Yes, this is for MIPS. On that processor an unexpected unaligned access cause a trap so this increased speed more than 3X. I was seeing that issue with the standard non-vec C version and a patched openssl so I can't yet reproduce directly with bssl.

Should the alignment be done inside poly1035.c as well instead? I was wondering why the structure was larger than needed. Alignment to a cache line size might benefit slightly so maybe both should use 64 bytes?

Patch-set: 1