Update patch set 2
Patch Set 2: Commit-Queue+1

Okay, wrote tests and stuff in the following commits. I couldn't find test vectors in the thread that hit these though. It seems the test vectors in the thread were assuming the field operations did not fully reduce things, but the eventual conclusion was that they should. It's not clear to me if, before this change, everything happened to work out anyway or not.

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Commit-Queue=+1