Update patch set 4
Patch Set 4:

Does it matter whether the input/output is in the Montgomery domain? I can't remember for sure, but I don't think it does. The important thing is that the output is in the same domain as the input.

Note in particular that the Montgomery Modular Inverse is a different computation from the Modular Inverse, so adding `_mont` to the name perhaps trades one confusion for another.

Patch-set: 4