Update patch set 1
Patch Set 1:

I think this is the best way to write this code, but it might be better to have more complete testing of the reduction logic with the "interesting" test vectors that haven't been added yet (if I understand your other PR). 

In particular, it is not so comforting that the original author (Vlad) of the code told us that this patch is flat-out wrong in the private email thread, although I guess we believe that he might be misunderstanding this or confusing it with the Almost-Montgomery logic in crypto/bn/asm.

Patch-set: 1
Label: Code-Review=0