Add C implementation of SHA-1 with AltiVec optimization.
This is a C implementation for POWER using AltiVec.  It is almost as fast as
the scalar-only assembly implementation for POWER/POWERPC family in OpenSSL
but it is easier to maintain and it allows error checking with tools like ASAN.

This is tested only for ppc64le.  It may nor may not work for other
platforms in the POWER/POWERPC familiy.

Change-Id: I790352e86d9c0cc4e1e57d11c5a0aa5b0780ca6b