Update patch set 4
Patch Set 4:

OK, right, sorry. I do think this change makes sense.

Again, the thing I find unsettling is that I don't understand why the code wasn't written this way in the first place. The (0, 0, _) representation is weird so there must have been some motivation for using it. The Intel people said the weird representation was used so that more code could be shared, but with this PR it seems like that's not the case. Unless they are referring to the 4-way SIMD code that is in OpenSSL but not in BoringSSL or to the C reference implementation. Don't know, but it makes me nervous.

Patch-set: 4
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0