Add a helper function for parsing extensions blocks.
TLS 1.3 adds a number of places with extensions blocks that don't easily
fit into our callbacks. To trim some of the boilerplate switch-cases,
add a helper function to pull fields out of this structure. Draft 18, in
particular, went nuts with the Certificate message.

For now, it rejects unknown extensions, but it will probably grow an
allow_unknown parameter for NewSessionTicket.

Change-Id: Iea8bf8ab216270c081dd63e79aaad9ec73b3b550