Update patch set 17
Patch Set 17:

(9 comments)

I didn't look at the rest of the CL since you'll probably end up changing the server half here:

TLS 1.3 server handshake should never touch the handshake transcript. A stateless HelloRetryRequest will not have that available. While we don't support that now, we will eventually, so pick implementation strategies compatible with it. This is important because if the implementation doesn't work out, we need to find this out now, not after the spec is done.

But it's fine. Instead:

1. Decide on the PSK you were using and everything. When parsing, make a note of the binder length and the binder you are interested in (spit out a CBS), but don't check it yet.

[Skip this for now: 2. If doing stateless CH2 rather than CH1, clone the [recovered] ssl->s3->handshake_hash. This will give you CH1 || HRR for CH2.]

3. EVP_DigestUpdate the ClientHello header. This will require a TLS/DTLS switch due to the different handshake headers. This is annoying, but it's what we expected to have to do when this feature was designed.

4. EVP_DigestUpdate the ClientHello prefix.

5. Compute the verify_data from that.

6. Check it against the binder you are interested. If a mismatch, abort the connection.

Patch-set: 17