Update patch set 3
Patch Set 3:

Hello, is there any update on the getrandom(2) test? Looks like Google hasn't imploded yet, so that's a good sign ;)

I updated this patch to use the dummy buffer in the initial test as well like the test patch does, and rebased on master.

But now that I think about it, the fact that the test patch that was merged used the GRND_NONBLOCK flag seems to defeat the purpose of testing whether the blocking when the entropy pools are not initialized (e.g. during boot) causes any problem, which IIRC was one of the reasons for having a test phase before merging the actual getrandom(2) support, am I misremembering this?

Anyway, hope all is going well :)

Patch-set: 3