Update patch set 1
Patch Set 1:

> Interesting. Is there any specific reason for this? Did it break clients? Servers? It would be nice to have an official IANA code point for this (TLSv1.2+RSA-PSS) at some point, don't know if anyone else would be interested.
> 
> (sorry for the random wall of questions, feel free to ignore if you want)

I'm not sure what you meant by official IANA code point. These code points come from TLS 1.3 which defines the same code point for both TLS 1.2 and TLS 1.3. The new SignatureScheme arrangement is backportable.
https://tlswg.github.io/tls13-spec/#rfc.section.4.2.3

I suppose they are not strictly official yet, but they're intended to be when 1.3 is finalized.

Sufficiently old versions of NSS are buggy and intolerant to new signature algorithms in a very bizarre way. It's not common and I don't think it will be fatal to TLS 1.3, but we're backing it out Chrome 55 to give affected sites a little more time to update. (It's not prominent in the UI, but this CL is not targeting BoringSSL's master branch.)

Patch-set: 1