Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Aside: Setters like this are wrong, but that ship has long sailed.
>  > There is a ton of complexity because all these objects have empty
>  > states and are piecemeal settable. The right API is getter-only
>  > immutable keys with some way to construct them fully-formed.
> 
> For the record, I would be very happy if I had an RSA constructor that simply takes non-null n, e and d without having a separate setter, that's what I end up doing anyway with RSA_new+RSA_set0_key.
> 
> Would that be a more acceptable patch? I can also look into getting it merged into OpenSSL as well.

The trouble is not that OpenSSL's new API is bad (though it is). It's that, while we are pretending to be 1.0.2, adding 1.1.0 APIs risks lots of compatibility problems. In some cases, we needed to add them ahead of time anyway (the SSL stack slow getting opaquified, or the refcount stuff), but for random structs we aren't in a hurry to hide like RSA, I'd prefer we do it in a single 1.1.0 version flip.

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0