Update patch set 4
Patch Set 4:

OCSP responses are large enough that duplicating it across every socket is actually going to be meaningful increase in per-connection memory usage. But we have CRYPTO_BUFFER now, so if you store it as one of those, the SSL_CTX => SSL transition can be ref-count bump.

Patch-set: 4
Label: Code-Review=0