Update patch set 5
Patch Set 5:

> > Is the motivation for this that staples change across time?
 > That's
 > > certainly true, but it might suggest a locked update of the
 > > SSL_CTX. (In general I'd like to reduce the amount of memory that
 > > SSL* takes up.)
 > 
 > The motivation is that different SSLs generated from a single
 > SSL_CTX might have different certificates and require different
 > staples. The idea is that both certificate and staple can be set at
 > cert_cb time.
 > 
 > In case of async sockets, different SSLs created from the same
 > SSL_CTX, with different certificates, might end up getting the
 > wrong staple (e.g. if cert_cb for both SSLs is executed one after
 > the other, but nothing is actually sent to clients yet, then
 > they'll both get the staple of the last one that called
 > SSL_CTX_set_ocsp_response).
 > 
 > The same applies to SCTs btw (and I'll need to look into a patch
 > for that as well).
 > 
 > Unfortunately in all these cases SSLs need to maintain their own
 > configuration, unless you assume that for every SSL you create a
 > new SSL_CTX (which would make SSL_CTX kind of useless IMO).
 > 
 > But anyway, if SSL_CTX_set_ocsp_response is used, then one can
 > probably assume SSL_set_ocsp_response won't actually be needed, so
 > if CRYPTO_BUFFER is used like David suggested the memory penalty
 > shouldn't be that big I think.

It's still the case that we have a ton of loose change hanging off the SSL* that adds up. I'm expecting we'll eventually do a copy-on-write thing, at which point all these APIs you're adding will become a performance cliff (i.e. as soon as you call it, sockets suddenly use a bunch of memory).

Are you aware of SSL_set_SSL_CTX? That's the way one usually does certificate swaps in OpenSSL.

(The session timeout thing is somewhat different because all session cache is keyed off of the initial SSL_CTX. I'd like to fix that, but would need to figure out what the consequences are for existing consumers.)

Patch-set: 5