Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > It's still the case that we have a ton of loose change hanging off
>  > the SSL* that adds up. I'm expecting we'll eventually do a
>  > copy-on-write thing, at which point all these APIs you're adding
>  > will become a performance cliff (i.e. as soon as you call it,
>  > sockets suddenly use a bunch of memory).
> 
> I wonder if it would make sense to have something like SSL_HANDSHAKE but allocated on SSL_new() and freed after the handshake terminates?
> 
> Once the connection is established all these configuration values shouldn't be needed anymore in the majority of cases.

Hrm. That's a thought. Unfortunately there's this awful SSL_clear API which means "reset everything for a fresh connection but keep the configuration". I tried a while back but was not able to get rid of it. :-/

There is also renegotiation, but we've removed renego as a server, so we could drop it as a server or as a client when renego is forbidden.

Though the motivation then was code cleanup. It hadn't occur to me that it would also solve our per-SSL config problem. Perhaps it's worth a second look.

> But anyway, I'm not sure I understand this "copy-on-write" thing, but isn't it the case already (for ocsp_response anyway) that allocation only happens "on write" (that is at SSL_set_ocsp_response time) given CRYPTO_BUFFER use?

The thought is all the random fields trickling in would be on an SSL_CONFIG struct that just points to the SSL_CTX one. Once you mutate any of them, we'd punt you to the more memory-intensive path and copy-on-write the whole thing.

>  > Are you aware of SSL_set_SSL_CTX? That's the way one usually does
>  > certificate swaps in OpenSSL.
> 
> Well yeah, but again, I'd need to maintain an SSL_CTX for each SSL, which I'm guessing would consume more memory, no?

Typically servers service quite a few more connections than hosts... :-)

Patch-set: 5