RFC: Refactor SSL configuration into SSL_CONFIG
THIS IS ONLY A PROOF-OF-CONCEPT.

TODO:
* Initial review to validate design
* Move more SSL config into SSL_CONFIG (OCSP response, SCT list, ...)
* Write documentation for internal functions
* ...

This is an attempt to move the SSL configuration shared by both SSL_CTX and
SSL into a common object, SSL_CONFIG. The object can be shared between SSL_CTX
and SSLs on SSL initialization. On configuration changes applied on the SSL,
the SSL_CONFIG is duplicated.

This was previously discussed in
https://boringssl-review.googlesource.com/#/c/12660/

Change-Id: I4fbdde95fc7d849e6f6062c6bbd0c18220c558dd