Work around language and compiler bug in memcpy, etc.
Most C standard library functions are undefined if passed NULL, even
when the corresponding length is zero. This gives them (and, in turn,
all functions which call them) surprising behavior on empty arrays.
Some compilers will miscompile code due to this rule. See also
https://www.imperialviolet.org/2016/06/26/nonnull.html

Workaround this in base.h. Note that because this depends on
BORINGSSL_IMPLEMENTATION to avoid injecting it into all our consumers,
external build definitions may need to be tweaked to pick this up.

BUG=23

Change-Id: Iad4e8562b1e22e83420d69cadf9fa8d9e8e4c218