Update patch set 3
Patch Set 3:

If you stop exporting `BN_mod_exp_mont_consttime` then people will still be able to reach it through the use of the BN_FLG_CONSTTIME flag, so the similar reduction logic would need to be moved up to be ahead of the call to `BN_mod_exp_mont_consttime` in `BN_mod_exp_mont`.

However, I think it is a bad idea to have the constant-time variant only available through the use of BN_FLG_CONSTTIME. Instead OpenSSL and BoringSSL should be deprecating the use of the BN_FLG_CONSTTIME flag.

So, I think that the API needs to be redesigned so that there is a public/exported function that does exponentiation on secret exponents in a really constant-time way that doesn't require reducing inputs. But, it's probably better to leave the specifics of the design to whoever can get BoringSSL/OpenSSL/LibreSSL developers to agree on such a new API.

Patch-set: 3