Update patch set 9
Abandoned

It seems we more-or-less agree that BN_FLG_CONSTTIME is bad and so we should just get rid of that. This patch may be useful for that, or maybe not. Regardless, probably not worth merging this, though it might be useful to refer to this if/when `BN_FLG_CONSTTIME` is removed.

Patch-set: 9
Status: abandoned
Tag: autogenerated:gerrit:abandon