Update patch set 3
Abandoned

In general I still think this is a good idea, but the reason why requires a lot of explanation, and maybe people won't feel that explanation justifies changing the default behavior. So let's just let the code be, at least for now.

Patch-set: 3
Status: abandoned
Tag: autogenerated:gerrit:abandon