Update patch set 1
Abandoned

IIRC, I originally wrote this test because I wanted to point out some dead code in one or more of the inversion functions. A better strategy for getting rid of that dead code would be to demonstrate mathematically that it is unneeded (if that is indeed the case). But, I'm planning instead to use a different algorithm altogether where it is obvious that no such reductions are necessary.

When I write the code for the new algorithm, I'll create more interesting test vectors that are data-driven.

So, we don't need to do this.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon